Removing duplicated factory
