+++ app/services/cms/providers/strapi/factories/email_component_factory.rb
@@ -3,27 +3,27 @@ module Cms
     module Strapi
       module Factories
         module EmailComponentFactory
-          def self.process_component(strapi_data)
-            component_name = strapi_data[:__component]
+          include BaseFactory
+          def self.generate_component(component_name, strapi_data)
             case component_name
-            when "email-content.text"
+            when "text"
               EmailComponents::Text.new(
                 blocks: strapi_data[:textContent]
               )
-            when "email-content.cta"
+            when "cta"
               EmailComponents::Cta.new(
                 text: strapi_data[:text],
                 link: strapi_data[:link]
               )
-            when "email-content.course-list"
+            when "course-list"
               EmailComponents::CourseList.new(
                 section_title: strapi_data[:sectionTitle],
                 remove_on_match: strapi_data[:removeOnMatch],
                 courses: strapi_data[:courses].map do |course|
                   EmailComponents::Course.new(
                     activity_code: course[:activityCode],
-                    substitute: strapi_data[:substitute],
-                    display_name: strapi_data[:displayName]
+                    substitute: course[:substitute],
+                    display_name: course[:displayName]
                   )
                 end
               )
